pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "test"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "aiohttp>=3.13.2",
    "bs4>=0.0.2",
    "curl-cffi>=0.13.0",
    "emoji>=2.15.0",
    "emojiswitch>=0.0.3",
    "fuzzywuzzy>=0.18.0",
    "httpx[socks]>=0.28.1",
    "ipykernel>=6.31.0",
    "ipython>=8.18.1",
    "jieba>=0.42.1",
    "matplotlib>=3.9.4",
    "meme-generator>=0.2.2",
    "msgpack>=1.1.2",
    "mypy>=1.19.1",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-memes>=0.8.1",
    "nonebot-plugin-orm[aiosqlite,default,sqlite]>=0.8.2",
    "nonebot2[fastapi,httpx]>=2.4.3",
    "openai>=2.6.1",
    "pip>=25.3",
    "psutil>=7.1.2",
    "pyppeteer>=2.0.0",
    "ts3>=2.0.0b2",
    "types-tqdm>=4.67.0.20250809",
    "wordcloud>=1.9.4",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = []
builtin_plugins = []

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
